#AGSMT-67: added writeTag functionality

+++ src/org/geo2tag/ashamapsexample/SettingsForm.java
@@ -18,19 +18,17 @@ public class SettingsForm extends Form implements CommandListener {
 	private static final String SETTINGS_FORM_HEADER = "Requests settings";
 	private static final String CHANNELS_CHOICE_GROUP = "Select channel to display";
 
-	private static final String BACK_COMMAND_TEXT = "Back";
 	
 	private TextField m_radiusEdit = new TextField(RADIUS_EDIT_TEXT, Settings.DEFAULT_RADIUS+"", 10, TextField.ANY);
 
 	
 	private ChoiceGroup m_channelsChoiceGroup = new ChoiceGroup(CHANNELS_CHOICE_GROUP, Choice.EXCLUSIVE);
 
-	private static final Command BACK_COMMAND = new Command(BACK_COMMAND_TEXT, Command.HELP, 0);
 
 	public SettingsForm() {
 		super(SETTINGS_FORM_HEADER);
 		
-		addCommand(BACK_COMMAND);
+		addCommand(Settings.BACK_COMMAND);
 
 		append(m_radiusEdit);
 		append(m_channelsChoiceGroup);
@@ -54,6 +52,8 @@ public class SettingsForm extends Form implements CommandListener {
 	}
 	
 	public void setChannels(Vector channels){
+		if (channels == null) return;
+
 		m_channelsChoiceGroup.deleteAll();
 		m_channelsChoiceGroup.append(Settings.SHOW_ALL_CHANNELS, null);
 		for (int i=0; i<channels.size(); i++){
